pyproject.toml
[project]
name = "osp-chatbot"
version = "0.0.1"
description = "BIP Haaga Helia 2025 - GenAI Project"
authors = [
    {name = "Paul Zoppoth"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "sentence-transformers (==2.2.2)",
    "torch (>=2.0.0)",
    "transformers (>=4.6.0)",
    "huggingface-hub (==0.16.4)",
    "langchain (==0.0.329)",
    "faiss-cpu (==1.7.4)",
    "python-dotenv (==1.0.0)",
    "tiktoken (==0.5.1)",
    "openai (==0.27.10)",
    "fireworks (==0.4.0)",
    "pypdf (==3.17.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
